remove unnecessary classes and do initial refactoring
