+++ src/main/java/com/infoshare/alpha/wwr/common/Facility.java
@@ -1,8 +1,4 @@
-package com.infoshare.alpha.wwr.domain.facilities.entity;
-
-import com.infoshare.alpha.wwr.common.Address;
-import com.infoshare.alpha.wwr.common.Service;
-import com.infoshare.alpha.wwr.common.Services;
+package com.infoshare.alpha.wwr.common;
 
 import java.util.ArrayList;
 import java.util.List;
