--- src/main/java/com/infoshare/alpha/wwr/di/DI.java
@@ -1,5 +1,0 @@
-//package com.infoshare.alpha.wwr.di;
-//
-//public interface DI {
-//
-//}
