+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/command/FacilityDeleteCommand.java
@@ -1,6 +1,6 @@
 package com.infoshare.alpha.wwr.domain.facilities.command;
 
-import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
+import com.infoshare.alpha.wwr.common.Facility;
 
 public class FacilityDeleteCommand extends FacilityCommand {
