+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/readmodel/FacilitiesReadModelDb.java
@@ -3,8 +3,8 @@
 //import java.util.List;
 //
 //import com.infoshare.alpha.wwr.common.Address;
-//import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
-//import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
+//import com.infoshare.alpha.wwr.common.Facilities;
+//import com.infoshare.alpha.wwr.common.Facility;
 //import com.infoshare.alpha.wwr.domain.facilities.query.FacilityPatientQuery;
 //import com.infoshare.alpha.wwr.domain.facilities.query.FacilityQuery;
 //
