+++ src/main/java/com/infoshare/alpha/wwr/utils/FacilitiesException.java
@@ -1,4 +1,4 @@
-package com.infoshare.alpha.wwr.domain.facilities.common;
+package com.infoshare.alpha.wwr.utils;
 
 public class FacilitiesException extends Throwable {
