+++ src/main/java/com/infoshare/alpha/wwr/utils/WwrController.java
@@ -6,9 +6,8 @@ import com.infoshare.alpha.wwr.domain.facilities.FacilitiesService;
 import com.infoshare.alpha.wwr.domain.facilities.command.FacilityAddCommand;
 import com.infoshare.alpha.wwr.domain.facilities.command.FacilityDeleteCommand;
 import com.infoshare.alpha.wwr.domain.facilities.command.FacilityEditCommand;
-import com.infoshare.alpha.wwr.domain.facilities.common.FacilitiesException;
-import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
-import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
+import com.infoshare.alpha.wwr.common.Facilities;
+import com.infoshare.alpha.wwr.common.Facility;
 import com.infoshare.alpha.wwr.domain.facilities.readmodel.FacilitiesReadModelDbRepository;
 import com.infoshare.alpha.wwr.domain.patients.PatientsService;
 import com.infoshare.alpha.wwr.domain.patients.entity.Patient;
